backend/requirements.txt
# ================================
# Health Routes Backend Dependencies
# ================================

# Core FastAPI Framework
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.20

# HTTP Client Libraries (for API calls)
httpx==0.27.2
requests==2.32.3

# Environment Configuration
python-dotenv==1.0.1

# Data Validation
pydantic==2.11.7

# Google Maps/Places APIs
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0

# Google Gemini AI API
google-api-core==2.19.0
googleapis-common-protos==1.63.1

# CORS Support
flask-cors==5.0.1

# Core Dependencies
anyio==4.6.2.post1
python-dateutil==2.9.0.post0
urllib3==2.2.1
certifi==2024.6.2

# Optional: Development and Testing
# pytest==8.3.5
# pytest-asyncio==0.26.0